Standardized on size_t over uint64_t and made read access const

+++ src/HTTP/Entity/withWriterAt.def.hpp
@@ -1,6 +1,6 @@
 template <typename Visitor>
 auto
-T::withWriterAt (Visitor && visitor, uint64_t position)
+T::withWriterAt (Visitor && visitor, size_t position)
 {
 	return std::visit
 	(
